update amrex submod (#876)
* update amrex submod

* clang-tidy fixes for AMREX_SPACEDIM

+++ amr-wind/core/Field.cpp
@@ -20,7 +20,9 @@ FieldInfo::FieldInfo(
     , m_ngrow(ngrow)
     , m_nstates(nstates)
     , m_floc(floc)
-    , m_bc_values(AMREX_SPACEDIM * 2, amrex::Vector<amrex::Real>(ncomp, 0.0))
+    , m_bc_values(
+          static_cast<long>(AMREX_SPACEDIM) * 2,
+          amrex::Vector<amrex::Real>(ncomp, 0.0))
     , m_bc_values_dview(static_cast<long>(ncomp) * AMREX_SPACEDIM * 2)
     , m_bcrec(ncomp)
     , m_bcrec_d(ncomp)
